Made it so you can make a profile with a profile picture and a bio. The profile pictures show up when viewing posts, and you can click on other people's profile to view their posts.

+++ public/admin_delete_post.php
@@ -22,9 +22,9 @@ if ($postId <= 0) {
 $imagePath = getPostImagePath($dbconn, $postId);
 
 if (deletePostById($dbconn, $postId) && $imagePath) {
-    $absoluteImagePath = __DIR__ . '/' . ltrim($imagePath, '/');
+    $absoluteImagePath = resolveStoredAssetFilesystemPath($imagePath);
 
-    if (is_file($absoluteImagePath)) {
+    if ($absoluteImagePath && is_file($absoluteImagePath)) {
         unlink($absoluteImagePath);
     }
 }
